+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 - 2.0.0
     - GBM estimators now use gradient-based subsampling
+    - Token Hashing Vectorizer custom hash functions
     - Change Token Hashing Vectorizer hashing function
     - Gradient Boost always uses mean base estimator
     - Move dummy estimators to the Extras package